Raise an exception when an after_deploy hook fail
This prevent the new deployment from being activated, but we keep the
files in place: a manual intervention might be required e.g. to restore
database at the state they where before we tried to create a new
deployment.

Previously, an error was silently ignored and could not be catch.