Strip leading `./` component from artifact files
When the artifact archive is created with `tar zcf filename.tgz .`, all
files path start with `./` but we do not want the application module to
believe that all files are under a `.` directory that can be used as the
deployment name.

Strip this `./` component to fix detection of a common root directory.