Use an AIO-agnostic path to Ruby for tasks
When running tasks with bolt, we can force an interpreter based on the
extension of the task we run.  With choria, the shebang is respected.
Bolt keep the user environment as is, and Choria makes sure the AIO path
is prepended before spawning the process in a clean environment, so this
is supposed to be backwards compatible and only less a PITA for users of
non-AIO Puppet.